python/requirements.txt
absl-py==1.0.0
aiohttp==3.8.0
aiosignal==1.2.0
alpha-vantage==2.3.1
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.0
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.1
beautifulsoup4==4.10.0
bitarray==1.2.2
bleach==4.1.0
blis==0.7.5
bs4==0.0.1
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.24
cytoolz==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
docopt==0.6.2
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.3.1
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.2.0
gast==0.4.0
glfw==2.4.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grip==4.5.2
grpcio==1.42.0
gym==0.21.0
h5py==3.6.0
hexbytes==0.2.2
httplib2==0.20.4
huggingface-hub==0.4.0
ibapi==9.76.1
idna==3.3
imageio==2.10.5
importlib-resources==5.4.0
ipfshttpclient==0.8.0a2
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jovian==0.2.41
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==12.0.0
lightgbm==3.3.2
lockfile==0.12.2
lru-dict==1.1.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
multiaddr==0.0.9
multidict==5.2.0
murmurhash==1.0.6
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
nmslib==2.1.1
notebook==6.4.5
numpy==1.21.4
oauthlib==3.1.1
onnxruntime==1.10.0
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.2
path-and-address==2.0.1
pathy==0.6.1
pbr==5.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.9.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.8.2
pyglet==1.5.21
Pygments==2.10.0
pyjnius==1.4.1
pyparsing==2.4.7
pyrsistent==0.18.0
pyserini==0.15.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
pytz==2021.3
pywin32==302
pywinpty==1.1.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
rdflib==6.0.2
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.1
scipy==1.7.2
semantic-version==2.8.5
Send2Trash==1.8.0
sentencepiece==0.1.96
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
solc-select==0.2.1
soupsieve==2.3.1
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sparky==0.0.1.dev0
SPARQLWrapper==1.8.5
sqlparse==0.4.2
srsly==2.4.2
stevedore==3.5.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.11.5
tomli==1.2.2
toolz==0.11.1
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
tweepy==4.4.0
typer==0.4.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uuid==1.30
varint==1.0.2
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wasabi==0.9.0
wcwidth==0.2.5
web3==5.24.0
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
